Remove GeoResources namespace
