+++ app/services/geo_discovery/document_builder/document_helper.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+module GeoDiscovery
+  class DocumentBuilder
+    class DocumentHelper
+      include Rails.application.routes.url_helpers
+      include ActionDispatch::Routing::PolymorphicRoutes
+    end
+  end
+end
