+++ spec/controllers/vector_resources_controller_spec.rb
@@ -205,10 +205,10 @@ RSpec.describe VectorResourcesController do
   describe "GET /vector_resources/:id/geoblacklight" do
     let(:user) { FactoryBot.create(:admin) }
     let(:vector_resource) { FactoryBot.create_for_repository(:vector_resource) }
-    let(:builder) { instance_double(GeoResources::GeoDiscovery::DocumentBuilder) }
+    let(:builder) { instance_double(GeoDiscovery::DocumentBuilder) }
 
     before do
-      allow(GeoResources::GeoDiscovery::DocumentBuilder).to receive(:new).and_return(builder)
+      allow(GeoDiscovery::DocumentBuilder).to receive(:new).and_return(builder)
     end
 
     context 'with a valid geoblacklight document' do
