+++ spec/services/geo_discovery/abstract_document_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 require 'rails_helper'
 
-describe GeoResources::GeoDiscovery::AbstractDocument do
+describe GeoDiscovery::AbstractDocument do
   subject(:abstract_document) { described_class.new }
 
   describe '#to_hash' do
