+++ spec/services/geo_discovery/document_builder/wxs_spec.rb
@@ -2,7 +2,7 @@
 require 'rails_helper'
 include ActionDispatch::TestProcess
 
-describe GeoResources::GeoDiscovery::DocumentBuilder::Wxs do
+describe GeoDiscovery::DocumentBuilder::Wxs do
   with_queue_adapter :inline
   subject(:wxs_builder) { described_class.new(decorator) }
